Split stores/helpers barrel into per-store shallow hooks
stores/helpers.ts eagerly imported all ten stores, so any shell module
(App, WindowFrame, MenuBar, MacDock) importing a single helper dragged
every store into the entry chunk. Keep the generic useStoreShallow in
helpers.ts and move each useXStoreShallow wrapper into its own store
file, then retarget all consumer imports.

Also move preloadIpodData/loadDefaultTracks into stores/ipodLibraryData
so main.tsx no longer statically imports the 2.6k-line iPod store just
to preload track data.
