+++ src/apps/chats/hooks/useChatRoom.ts
@@ -10,7 +10,7 @@ import { useChatsStore } from "@/stores/useChatsStore";
 import { useIsRyoAdmin } from "@/hooks/useIsRyoAdmin";
 import { toast } from "@/hooks/useToast";
 import { type ChatRoom, type ChatMessage } from "@/types/chat";
-import { useChatsStoreShallow } from "@/stores/helpers";
+import { useChatsStoreShallow } from "@/stores/useChatsStore";
 import { removeChatRoomById, upsertChatRoom } from "@/utils/chatRoomList";
 import { shouldNotifyForRoomMessage } from "@/utils/chatNotifications";
 import { showRoomMessageNotification } from "@/utils/chatNotificationDisplay";
