+++ src/apps/terminal/hooks/useTerminalLogic.ts
@@ -6,7 +6,7 @@ import {
 } from "@/apps/finder/hooks/useFileSystem";
 import { useThemeFlags } from "@/hooks/useThemeFlags";
 import { STORES, dbOperations } from "@/utils/indexedDB";
-import { useTerminalStoreShallow } from "@/stores/helpers";
+import { useTerminalStoreShallow } from "@/stores/useTerminalStore";
 import { useTerminalStore } from "@/stores/useTerminalStore";
 import { useLaunchApp } from "@/hooks/useLaunchApp";
 import { useAiChat } from "@/apps/chats/hooks/useAiChat";
