+++ src/components/layout/ExposeView.tsx
@@ -2,7 +2,7 @@ import { useEffect, useCallback, useMemo } from "react";
 import { useEventListener } from "@/hooks/useEventListener";
 import { usePrevious } from "@/hooks/useLatestRef";
 import { motion, AnimatePresence } from "motion/react";
-import { useAppStoreShallow } from "@/stores/helpers";
+import { useAppStoreShallow } from "@/stores/useAppStore";
 import { getAppIconPath } from "@/config/appRegistry";
 import { getTranslatedAppName } from "@/utils/i18n";
 import { ThemedIcon } from "@/components/shared/ThemedIcon";
