+++ src/components/layout/menu-bar/DefaultMenuItems.tsx
@@ -11,7 +11,7 @@ import {
 import { HelpDialog } from "@/components/dialogs/HelpDialog";
 import { AboutDialog } from "@/components/dialogs/AboutDialog";
 import { useLaunchApp } from "@/hooks/useLaunchApp";
-import { useAppStoreShallow } from "@/stores/helpers";
+import { useAppStoreShallow } from "@/stores/useAppStore";
 import { useUndoRedoStore } from "@/stores/useUndoRedoStore";
 import { ThemedIcon } from "@/components/shared/ThemedIcon";
 import { useTranslatedHelpItems } from "@/hooks/useTranslatedHelpItems";
