+++ src/components/layout/menu-bar/MacTopMenuBar.tsx
@@ -2,7 +2,8 @@ import type { ReactNode } from "react";
 import { Menubar } from "@/components/ui/menubar";
 import { AppleMenu } from "../AppleMenu";
 import { AppMenu } from "../AppMenu";
-import { useAppStoreShallow, useDisplaySettingsStoreShallow } from "@/stores/helpers";
+import { useAppStoreShallow } from "@/stores/useAppStore";
+import { useDisplaySettingsStoreShallow } from "@/stores/useDisplaySettingsStore";
 import { useThemeFlags } from "@/hooks/useThemeFlags";
 import { useIsPhone } from "@/hooks/useIsPhone";
 import { isTauri, isTauriWindows } from "@/utils/platform";
