+++ src/components/screensavers/ScreenSaverOverlay.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useReducer, useCallback, lazy, Suspense } from "react";
-import { useDisplaySettingsStoreShallow } from "@/stores/helpers";
+import { useDisplaySettingsStoreShallow } from "@/stores/useDisplaySettingsStore";
 import { motion, AnimatePresence } from "motion/react";
 import type { ScreenSaverType } from "./index";
 import { useTranslation } from "react-i18next";
