+++ src/main.tsx
@@ -5,7 +5,7 @@ import "./index.css";
 import { useThemeStore } from "./stores/useThemeStore";
 import { useLanguageStore } from "./stores/useLanguageStore";
 import { preloadFileSystemData } from "./stores/useFilesStore";
-import { preloadIpodData } from "./stores/useIpodStore";
+import { preloadIpodData } from "./stores/ipodLibraryData";
 import { initPrefetch } from "./utils/prefetch";
 import { initializeI18nForFirstPaint } from "./lib/i18n";
 import { primeReactResources } from "./lib/reactResources";
