+++ src/stores/useIpodStore.ts
@@ -1,5 +1,6 @@
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
+import { useStoreShallow } from "./helpers";
 import {
   LyricsAlignment,
   KoreanDisplay,
@@ -12,7 +13,7 @@ import {
 import { LyricLine } from "@/types/lyrics";
 import type { FuriganaSegment } from "@/utils/romanization";
 import { getAppPublicOrigin } from "@/utils/runtimeConfig";
-import { getCachedSongMetadata, listAllCachedSongMetadata } from "@/utils/songMetadataCache";
+import { getCachedSongMetadata } from "@/utils/songMetadataCache";
 import { ApiRequestError } from "@/api/core";
 import {
   clearSongCachedData,
@@ -28,6 +29,7 @@ import {
 } from "@/utils/youtubeMetadata";
 import { parseYouTubeVideoId } from "@/utils/youtubeUrl";
 import { emitCloudSyncDomainChange } from "@/utils/cloudSyncEvents";
+import { loadDefaultTracks } from "@/stores/ipodLibraryData";
 import { sortTracksLikeServerOrder } from "@/stores/ipodTrackOrder";
 import {
   hasFetchedTrackMetadataChanges,
@@ -35,7 +37,6 @@ import {
   resolveSyncedCoverColor,
   shouldUpdateTrackLyricsSource,
 } from "@/stores/ipodTrackMetadataSync";
-import { mapCatalogSongToTrack } from "@/stores/ipodCatalogTrackMapping";
 import {
   saveAppleMusicLibrary,
   saveAppleMusicPlaylistTracks,
@@ -353,83 +354,8 @@ interface IpodData {
   ipodMenuMode: boolean | null;
 }
 
-// ============================================================================
-// CACHING FOR iPod TRACKS
-// ============================================================================
-
-// In-memory cache for iPod tracks data
-let cachedIpodData: { tracks: Track[]; version: number } | null = null;
-let ipodDataPromise: Promise<{ tracks: Track[]; version: number }> | null = null;
-/** Only the latest load may write `cachedIpodData` (avoids stale force-refresh overwrites). */
-let ipodLoadGeneration = 0;
-
-/**
- * Preload iPod tracks data early (can be called before React mounts).
- * This starts fetching the JSON file without blocking.
- */
-export function preloadIpodData(): void {
-  if (cachedIpodData || ipodDataPromise) return;
-  loadDefaultTracks();
-}
-
-/**
- * Load default tracks from Redis song metadata cache.
- * @param forceRefresh - If true, bypasses cache and fetches fresh data (used by syncLibrary)
- */
-async function loadDefaultTracks(forceRefresh = false): Promise<{
-  tracks: Track[];
-  version: number;
-}> {
-  // Return cached data immediately if available (unless force refresh)
-  if (!forceRefresh && cachedIpodData) {
-    return cachedIpodData;
-  }
-  
-  // Return existing promise if fetch is in progress (deduplication)
-  // But not if we need a force refresh
-  if (!forceRefresh && ipodDataPromise) {
-    return ipodDataPromise;
-  }
-  
-  const thisGeneration = ++ipodLoadGeneration;
-
-  // Start new fetch
-  const fetchPromise = (async () => {
-    try {
-      // Load from Redis song metadata cache
-      // Only sync songs created by user "ryo" (the admin/curator)
-      const cachedSongs = await listAllCachedSongMetadata("ryo");
-      
-      console.log(`[iPod Store] Loaded ${cachedSongs.length} tracks from Redis cache (by ryo)`);
-      // Songs are already sorted by createdAt (newest first) from the API
-      const tracks: Track[] = cachedSongs.map(mapCatalogSongToTrack);
-      // Use the latest createdAt timestamp as version (or 1 if empty)
-      const version = cachedSongs.length > 0 
-        ? Math.max(...cachedSongs.map((s) => s.createdAt || 1))
-        : 1;
-      const payload = { tracks, version };
-      if (thisGeneration === ipodLoadGeneration) {
-        cachedIpodData = payload;
-        return payload;
-      }
-      // A newer load won the race; prefer latest cache so awaiters do not apply stale tracks.
-      return cachedIpodData ?? payload;
-    } catch (err) {
-      console.error("Failed to load tracks from cache", err);
-      return { tracks: [], version: 1 };
-    }
-  })();
-  
-  // Only set the shared promise for non-force-refresh requests
-  if (!forceRefresh) {
-    ipodDataPromise = fetchPromise;
-    fetchPromise.finally(() => {
-      ipodDataPromise = null;
-    });
-  }
-  
-  return fetchPromise;
-}
+// Track caching/preload lives in @/stores/ipodLibraryData so that main.tsx can
+// preload tracks without pulling this entire store into the entry chunk.
 
 const initialIpodData: IpodData = {
   tracks: [],
@@ -2637,4 +2563,10 @@ if (import.meta.hot) {
     };
     (data as { [HMR_KEY]?: Partial<IpodData> })[HMR_KEY] = snapshot;
   });
+}
+
+export function useIpodStoreShallow<T>(
+  selector: (state: ReturnType<typeof useIpodStore.getState>) => T
+): T {
+  return useStoreShallow(useIpodStore, selector);
 }
